requirements.txt
os
pandas
streamlit
dotenv
google-generativeai
pathlib